LayoutTests/fast/events/detached-svg-parent-window-events-expected.txt

This tests creating a disconnected SVG element with resize event handler. The event handler should not get dispatched unless the element is connected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS iframe.contentWindow.eval('throw "error"') threw exception error.
PASS didFireResize is false
PASS didFireOnError is false
PASS successfullyParsed is true

TEST COMPLETE
